requirements.txt
openai
streamlit
pydantic
python-dotenv
streamlit-mermaid
